psql_to_redshift/pyproject.toml
[tool.poetry]
name = "psql_to_redshift"
version = "0.1.0"
description = ""
authors = ["vasilis <vasilis.ntousis@legal-i.ch>"]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.22.3"
python-dotenv = "^0.20.0"
psycopg2-binary = "^2.9.3"
black = "^22.3.0"
boto3 = "^1.22.1"
ansible = "^5.7.0"
redshift-connector = "^2.0.906"
pyodbc = "^4.0.32"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
